feat: refactor code to use enums
- Migrate Enums to separate files
- Removed redundant Enum definitions
