+++ models.py
@@ -1,0 +1,6 @@
+from enum import Enum, auto
+
+
+class Models(Enum):
+    GPT_3_5_TURBO = "gpt-3.5-turbo"
+    LLAMA3 = "llama3"
